requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
python-dotenv
psycopg
flask-cors